+++ src/Core2D.Core/ViewModels/Layout/Box.cs
@@ -2,12 +2,12 @@
 
 public struct Box
 {
-    public decimal Left;
-    public decimal Top;
-    public decimal Right;
-    public decimal Bottom;
-    public decimal CenterX;
-    public decimal CenterY;
-    public decimal Width;
-    public decimal Height;
+    public double Left;
+    public double Top;
+    public double Right;
+    public double Bottom;
+    public double CenterX;
+    public double CenterY;
+    public double Width;
+    public double Height;
 }
