+++ src/Core2D.Core/ViewModels/Shapes/ArcShapeViewModel.cs
@@ -103,7 +103,7 @@ public partial class ArcShapeViewModel : BaseShapeViewModel
         _point4?.Bind(dataFlow, db, record);
     }
 
-    public override void Move(ISelection? selection, decimal dx, decimal dy)
+    public override void Move(ISelection? selection, double dx, double dy)
     {
         if (_point1 is null || _point2 is null || _point3 is null || _point4 is null)
         {
