+++ src/Core2D.Core/ViewModels/Shapes/PathShapeViewModel.cs
@@ -92,7 +92,7 @@ public partial class PathShapeViewModel : BaseShapeViewModel
         }
     }
 
-    public override void Move(ISelection? selection, decimal dx, decimal dy)
+    public override void Move(ISelection? selection, double dx, double dy)
     {
         foreach (var point in GetPathPoints())
         {
